ConectaVid's First Release (1.0)
The first ConectaVid release includes a PDF Printer, which prints a vaccination report into a "File" folder, in the user's main HDD root folder, a patient register, and a full-working login platform. It uses a MySQL server to store the database.
For the next release, I plan to complete the patient registration form and allow the registration of new users (such as admins or normal attendants).